Various fixes

+++ lib/utf8.h
@@ -60,7 +60,7 @@ char *u8_strchr(char *s, u_int32_t ch, int *charn);
 char *u8_memchr(char *s, u_int32_t ch, size_t sz, int *charn);
 
 /* count the number of characters in a UTF-8 string */
-int u8_strlen(const char *s);
+size_t u8_strlen(const char *s);
 
 int u8_is_locale_utf8(char *locale);
